server/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
turbopuffer
elevenlabs
librosa
numpy
scipy
huggingface-hub
